+++ source/extensions/omni.isaac.orbit/omni/isaac/orbit/envs/rl_env_cfg.py
@@ -7,6 +7,7 @@ from __future__ import annotations
 
 from dataclasses import MISSING
 
+from omni.isaac.orbit.command_generators import CommandGeneratorBaseCfg
 from omni.isaac.orbit.utils import configclass
 
 from .base_env_cfg import BaseEnvCfg
@@ -25,7 +26,7 @@ class RLEnvCfg(BaseEnvCfg):
     """Reward settings."""
     terminations: object = MISSING
     """Termination settings."""
-    randomization: object = MISSING
-    """Randomization settings."""
     curriculum: object = MISSING
     """Curriculum settings."""
+    commands: CommandGeneratorBaseCfg = MISSING
+    """Command generator settings."""